+++ src/main/java/net/litetex/capes/fabric/compat/ModMenuCompatibility.java
@@ -1,4 +1,4 @@
-package net.litetex.capes.fabric.compatibility;
+package net.litetex.capes.fabric.compat;
 
 import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;